% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyBars.R
\name{plotlyBars}
\alias{plotlyBars}
\title{accept only a single ggplot layer}
\usage{
plotlyBars(input, output, session, plot_reactive)
}
\arguments{
\item{plot_reactive}{A reactive returning a plotly object}
}
\description{
accept only a single ggplot layer
}
\details{
This module wraps a reactive returning
}
