% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyBars.R
\name{plotlyBarsUI}
\alias{plotlyBarsUI}
\title{The UI function}
\usage{
plotlyBarsUI(id, ...)
}
\arguments{
\item{id}{The namespace id}

\item{...}{Further arguments passed to \code{\link{plotlyOutput}}}
}
\description{
The UI function
}
\details{
See \url{https://github.com/andrewsali/plotlyBars}
}
