% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withBars.R
\name{withBars}
\alias{withBars}
\title{Utility functions to convert regular plotly shiny commands to module commands}
\usage{
withBars(plotly_output_expr)
}
\description{
Utility functions to convert regular plotly shiny commands to module commands
}
