% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withBars.R
\name{withBarsUI}
\alias{withBarsUI}
\title{Utility functions to convert regular plotly shiny commands to module commands}
\usage{
withBarsUI(plotly_ui_expr)
}
\description{
Utility functions to convert regular plotly shiny commands to module commands
}
