% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyBars.R
\name{plotlyBars}
\alias{plotlyBars}
\title{Module function which calls \code{\link{renderPlotly}} and ensures that loading animation is appropriately displayed / hidden.}
\usage{
plotlyBars(input, output, session, plot_reactive, ...)
}
\arguments{
\item{plot_reactive}{A reactive returning a plotly object}

\item{...}{Further options passed to \code{\link{renderPlotly}}}
}
\description{
Module function which calls \code{\link{renderPlotly}} and ensures that loading animation is appropriately displayed / hidden.
}
\details{
See \url{https://github.com/andrewsali/plotlyBars}
}
